// types/chat.ts
export enum BotConversationMessageSenderType {
	Facilitator = "FACILITATOR",
	System = "SYSTEM",
	User = "USER",
	AI = "AI",
}

export enum BotConversationMessageType {
	Reflection_Selection_Facilitator_Message = "REFLECTION_SELECTION_FACILITATOR_MESSAGE",
	Tool_Selection_Facilitator_Message = "TOOL_SELECTION_FACILITATOR_MESSAGE",
	Notebook_Block_User_Image_Message = "NOTEBOOK_BLOCK_USER_IMAGE_MESSAGE",
	General_Intermediate_Message = "GENERAL_INTERMEDIATE_MESSAGE",
	AI_Generated_Search_Query = "AI_GENERATED_SEARCH_QUERY",
	Plan_Next_Step_Response = "PLAN_NEXT_STEP_RESPONSE",
	Complete_Model_Response = "COMPLETE_MODEL_RESPONSE",
	Tool_Selection_Response = "TOOL_SELECTION_RESPONSE",
	Notebook_Block_Message = "NOTEBOOK_BLOCK_MESSAGE",
	Refining_Plan_Message = "REFINING_PLAN_MESSAGE",
	Complete_Tool_Message = "COMPLETE_TOOL_MESSAGE",
	Relevant_Urls_Message = "RELEVANT_URLS_MESSAGE",
	Start_Model_Response = "START_MODEL_RESPONSE",
	First_System_Message = "FIRST_SYSTEM_MESSAGE",
	Start_Tool_Message = "START_TOOL_MESSAGE",
	Reflection_Message = "REFLECTION_MESSAGE",
	Sources_Message = "SOURCES_MESSAGE",
	Tool_Response = "TOOL_RESPONSE",
	User_Message = "USER_MESSAGE",
	AI_Response = "AI_RESPONSE",
}

export enum BotConversationMessageStatus {
	InProgress = "IN_PROGRESS",
	Complete = "COMPLETE",
	Error = "ERROR",
}

export type BotConversationMessageUuid = Tagged<
	string,
	"BotConversationMessageUuid"
>;
export type BotConversationMessageId = Tagged<
	number,
	"BotConversationMessageId"
>;
export type ParallelConversationId = Tagged<number, "ParallelConversationId">;

export type BotConversationMessage = {
	parallel_conversation_id: ParallelConversationId | null;
	message_status?: BotConversationMessageStatus;
	bot_conversation: { id: BotConversationId };
	sub_conversation_uuids: (string | null)[];
	block: { uuid: NotebookBlockUuid } | null;
	show_as_intermediate_step?: boolean;
	uuid: BotConversationMessageUuid;
	sources: SourceForUser[] | null;
	order_by_timestamp_ms: number;
	id: BotConversationMessageId;
	toggle_text?: string | null;
	thumbs_up: boolean | null;
	created_at: ISODateString;
	updated_at: ISODateString;
	visible_to_model: boolean;
	visible_to_user: boolean;
	text: string | null;
	transient: boolean;
	archived: boolean;
	sender: {
		sender_type: BotConversationMessageSenderType;
		sender_info: BetterbrainUser | null;
	};
} & (
	| {
			message_type: BotConversationMessageType.Tool_Response;
			json: ToolResponseJson;
	  }
	| {
			message_type: BotConversationMessageType.Tool_Selection_Response;
			json: ToolSelectionResponseJson;
	  }
	| {
			message_type: BotConversationMessageType.AI_Generated_Search_Query;
			json: AIGeneratedSearchQueryJson;
	  }
	| {
			message_type: BotConversationMessageType.Reflection_Message;
			json: ReflectionMessageJson;
	  }
	| {
			message_type: BotConversationMessageType.Plan_Next_Step_Response;
			json: PlanNextStepJson;
	  }
	| {
			message_type: BotConversationMessageType.Notebook_Block_User_Image_Message;
			block: { uuid: NotebookBlockUuid };
			json: null;
	  }
	| {
			message_type: BotConversationMessageType.General_Intermediate_Message;
			json: GeneralIntermediateMessageJson;
	  }
	| {
			message_type: BotConversationMessageType.Relevant_Urls_Message;
			json: RelevantUrlsMessageJson;
	  }
	| {
			message_type:
				| BotConversationMessageType.Reflection_Selection_Facilitator_Message
				| BotConversationMessageType.Tool_Selection_Facilitator_Message
				| BotConversationMessageType.Complete_Model_Response
				| BotConversationMessageType.Notebook_Block_Message
				| BotConversationMessageType.Refining_Plan_Message
				| BotConversationMessageType.Complete_Tool_Message
				| BotConversationMessageType.Start_Model_Response
				| BotConversationMessageType.First_System_Message
				| BotConversationMessageType.Start_Model_Response
				| BotConversationMessageType.Start_Tool_Message
				| BotConversationMessageType.Sources_Message
				| BotConversationMessageType.User_Message
				| BotConversationMessageType.AI_Response;
			json: null | undefined;
	  }
);
// ...
export type SourceForUser =
	| StandardDocumentSourcesForUser
	| ModeDefinitionSourceForUser
	| GoogleDriveSourcesForUser
	| DataSchemaSourceForUser
	| ModeQuerySourceForUser
	| WebsitesSourcesForUser
	| AffinitySourceForUser
	| AirtableSourceForUser
	| SQLQuerySourceForUser
	| SlackSourceForUser
	| PDFSourceForUser
	| WebSourceForUser;

type WebsitesSourcesForUser = {
	source_type: SourceForUserType.Website;
	data_key: string;
	source_info: {
		websites: Array<WebsiteSource>;
		type: never;
	};
};

export type WebsiteSource = {
	identifier: SourceID;
	type: "WEBSITE";
	link?: string;
};

type StandardDocumentSourcesForUser = {
	source_type: SourceForUserType.StandardDocument;
	data_key: string;
	source_info:
		| {
				documents: Array<MinimalStandardDocumentSourceValues>;
				type: StandardDocumentSourceType.Minimal;
		  }
		| {
				documents: VerboseStandardDocumentSourceValues;
				type: StandardDocumentSourceType.Verbose;
		  };
};
// ...
export type AirtableSourceForUser = {
	source_type: SourceForUserType.Airtable;
	data_key: string;
	source_info: {
		records: Array<AirtableRecord>;
		type: never;
	};
};
// ...
export type AffinitySourceForUser = {
	source_type: SourceForUserType.Affinity;
	data_key: string;
	source_info:
		| {
				type: AffinitySourceType.AffinityOrganization;
				organizations: Array<AffinityOrganization>;
		  }
		| {
				type: AffinitySourceType.AffinityPerson;
				persons: Array<AffinityPerson>;
		  }
		| {
				type: AffinitySourceType.AffinityNote;
				notes: Array<AffinityNote>;
		  }
		| {
				type: AffinitySourceType.AffinityIntroductionsMadeBy;
				introductions: Array<AffinityIntroductionsMadeBy>;
		  }
		| {
				type: AffinitySourceType.AffinityIntroductionsMadeTo;
				introductions: Array<AffinityIntroductionsMadeTo>;
		  };
};
// ...
export type ModeDefinitionSourceForUser = {
	source_type: SourceForUserType.ModeDefinition;
	data_key: string;
	source_info: {
		mode_definitions: Array<ModeDefinitionType>;
		type: never;
	};
};

export type ModeDefinitionType = {
	connection_type: DatabaseConnectionType;
	mode_definition_id: string;
	connection_id: number;
	description: string;
	relevance?: number;
	id?: SourceID;
	link?: string;
	query: string;
	name: string;
};

export type ModeQuerySourceForUser = {
	source_type: SourceForUserType.ModeQuery;
	data_key: string;
	source_info: {
		mode_queries: Array<ModeQueryType>;
		type: never;
	};
};
// ...
export type GoogleDriveSourcesForUser = {
	source_type: SourceForUserType.GoogleDrive;
	data_key: string;
	source_info:
		| {
				documents: GoogleDriveVerboseSource;
				type: GoogleDriveSourceType.Verbose;
		  }
		| {
				documents: Array<GoogleDriveMinimalSource>;
				type: GoogleDriveSourceType.Minimal;
		  };
};

export type SQLQuerySourceForUser = {
	source_type: SourceForUserType.SqlQuery;
	data_key: string;
	source_info: {
		queries: Array<SQLQueryType>;
		type: never;
	};
};
// ...
export type DataSchemaSourceForUser = {
	source_type: SourceForUserType.DataSchema;
	data_key: string;
	source_info: {
		entities: Array<DataSchemaEntity>;
		type: never;
	};
};
// ...
export type SlackSourceForUser = {
	source_type: SourceForUserType.Slack;
	data_key: string;
	source_info: {
		conversations: Array<SlackConversation>;
		type: never;
	};
};

export type SlackConversation = {
	channel_name?: string;
	channel_id: string;
	relevance?: number;
	id?: SourceID;
	link?: string;
	url: string;
	messages: Array<{
		sent_at: ISODateString;
		message_id: string;
		sender: string;
		text: string;
		replies: Array<{
			sent_at: ISODateString;
			message_id: string;
			sender: string;
			text: string;
		}>;
	}>;
};

export type WebSourceForUser = {
	source_type: SourceForUserType.Web;
	data_key: string;
	source_info: {
		website_snippets: Array<WebsiteSnippet>;
		type: never;
	};
};

export type WebsiteSnippet = {
	document_type: WebsiteDocumentType;
	description: string;
	website_id: number;
	relevance?: number;
	author: string;
	link?: string;
	title: string;
	id: SourceID;
	text: string;
	url: string;
};

enum WebsiteDocumentType {
	WebsiteContent = "WEBSITE_CONTENT",
	Description = "DESCRIPTION",
	Author = "AUTHOR",
	Title = "TITLE",
}

export type PDFSourceForUser = {
	source_type: SourceForUserType.Pdf;
	data_key: string;
	source_info: {
		pdf_snippets: Array<PdfSnippet>;
		type: never;
	};
};
// ...
export type ToolResponseJson = {
	tool:
		| {
				outputs: AskClarifyingQuestionOutputs;
				name: ToolType.AskClarifyingQuestion;
		  }
		| {
				name: ToolType.AnswerQuestionUsingContext;
				outputs: AnswerQuestionUsingContextOutputs;
		  }
		| {
				name: ToolType.PerformActionInNotebook;
				outputs: PerformActionInNotebookOutputs;
		  };
};

export type ToolSelectionResponseJson = {
	reasoning: string;
	tool:
		| {
				name: ToolType.AnswerQuestionUsingContext;
				inputs: AnswerQuestionUsingContextInputs;
		  }
		| {
				name: ToolType.AskClarifyingQuestion;
				inputs: AskClarifyingQuestionInputs;
		  }
		| {
				name: ToolType.PerformActionInNotebook;
				inputs: PerformActionInNotebookInputs;
		  }
		| {
				name: ToolType.ReturnNormalTextResponse;
				inputs: ReturnNormalTextResponseInNotebookInputs;
		  }
		| {
				name: ToolType.Planner;
				inputs: PlannerInputs;
		  };
};

type GeneralIntermediateMessageJson = {
	[url: string]: {
		does_context_contain_answer: boolean;
		sources: Array<WebsiteSource>;
		answer: string | null;
	};
};

type RelevantUrlsMessageJson = Array<string>;

export type AIGeneratedSearchQueryJson = {
	search_query: string;
};

export type ReflectionMessageJson = {
	is_task_complete: boolean;
	reflection: string;
	next_steps: string;
	reasoning: string;
};

export type PlanNextStepJson = {
	action_type: string;
	reasoning: string;
};

export type AnswerQuestionUsingContextInputs = {
	question: string;
};

export type AskClarifyingQuestionInputs = {
	question: string;
};

export type ReturnNormalTextResponseInNotebookInputs = {
	response: string;
};

export type PlannerInputs = {
	action_type: PlannerActionType;
	tasks: PlannerTask[];
};
// ...
type PerformActionInNotebookInputs = {
	connection_type:
		| DatabaseConnectionType.Postgres
		| DatabaseConnectionType.Snowflake
		| DatabaseConnectionType.BigQuery
		| DatabaseConnectionType.Slack;
	block_type: BlockType.Sql | BlockType.Python;
	action: PerformActionInNotebookActionType;
	sql_block_source_type: SqlBlockSourceType;
	write_variable: string;
	connection_id: number;
	block_uuid: string;
	question: string;
	execute: boolean;
};

type AskClarifyingQuestionOutputs = null;

type AnswerQuestionUsingContextOutputs = {
	sources: SourceForUser[];
	answer: string;
};

type PerformActionInNotebookOutputs = {
	execution_results_preview?: string;
	tool_thoughts?: string[];
	status?: string;
	error?: string;
};

export enum ToolType {
	AnswerQuestionUsingContext = "ANSWER_QUESTION_USING_CONTEXT",
	ReturnNormalTextResponse = "RETURN_NORMAL_TEXT_RESPONSE",
	PerformActionInNotebook = "PERFORM_ACTION_IN_NOTEBOOK",
	AskClarifyingQuestion = "ASK_CLARIFYING_QUESTION",
	Planner = "PLANNER",
}